Add files via upload
This represents the start minimal implementation guide. The guide has only a single page and one 'external' structure definition

There are several files in the 'examples' folder. These are not yet used by the IG in this commit. They were generated in Simplifier DSTU3
and will eventually be used in the PHD IG.

There are also several structure definitions in the resources file. Only two of them are used by the IG at this time. They are the
PhdImplementationGuide.xml and the PhdDeviceStructuredefinition.xml file. The latter is part of the PHD profile. The remaining xml
files are also part of the PHD profile ported from Simplifier/Forge and will need to be modified to support R4 3.6.0. But they are
not part of the minimal guide.

The ig.json and PhdImplementationGuide.xml files determine what resources are part of the published guide. At this commit, these
files include only the single page and PhdDeviceStructuredefinition resource.

When built, the index.html generated in /output will contain a single page showing the PhdDeviceStructuredefinition as a Table
(testingsome liquid syntax),  difference, snapshot, and 'dict' (dictionary).