Added string enumeration profile, trying to add new page. Clean up struct def text
The new page to add has been the profiles. This exposed some text in the struct defs
that could be better said. I may have to move all my pages to the  _includes directory
or some other directory in the pages directory to access them. Still not clear how to
make a page tree or access pages from a drop down menu