Addressing slicing on the components. Made new values sets for MDCObject. New slicing attempt
After discussion with LLoyd McKenzie put the slicing on value/code and put the code binding in the
code element TOP not the code.coding.code. To do so required MDC Object value sets for Quantities,
CodeableCOncpets, Ranges, and Strings. THe compounds went to the MDC Non Object and all ASN1 went to
one of the two ASN1 value sets. IN the latter case, two separate value sets are not even needed.